Change package names to reference VividCortex's fork
Notice that the plugin used to compile .proto files is changed also,
cause there's no option to force a new import path. Google developers
fixed the code.google.com URL within the tool. (The change in package
name was done only in code; documentation still reflects the locations
stated in the original repository.)

Code generated with the tool in this repository imports packages in
github.com/VividCortex/goprotobuf. A slightly different tool calls for a
slightly different name, to avoid confusion and preserve the semantics
of the original name. So I renamed the binary to protoc-gen-govc, with
the final vc standing for VividCortex, not surprisingly. To compile with
this version, you should use protoc --govc_out=<dir> file.proto,
provided protoc-gen-govc is in your path.

Closes #1.